+++ ot_simple_rest_job_proxy/views/__init__.py
@@ -1,0 +1,1 @@
+from .views import checkjob, makejob, getresult