+++ tests/test_dispatcher/test_otl_job_handler.py
@@ -81,14 +81,15 @@ class TestOtlJobHandler(BaseApiTest):
         self.assertEqual(len(node_jobs), 3)
         for node_job in node_jobs:
             self.assertEqual(
-                node_job.result.calculated, True
+                node_job.result.status, 'CALCULATED'
             )
             self.assertEqual(node_job.status, NodeJobStatus.FINISHED)
 
     def tearDown(self):
-        self.dispatcher_process.kill()
         self.spark_computing_node.kill()
         self.eep_computing_node.kill()
         self.pp_computing_node.kill()
+        self.dispatcher_process.kill()
+