+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2022-04-25
+### Added
+- Proxy plugin for ot_simple_rest
+- Result dataframe managing 
+- Cancel job endpoint
+### Fixed
+- Mock computing node accepts canceled status
+- Fix disappearing nodes from computing node pool when dispatcher restarts
+- Change job status to FAILED when syntaz error
+### Changed
+- Otlang version to 1.1.0
+
 ## [0.1.0] - 2022-03-29
 ### Added
 - Changelog.md.